+++ src/aks-preview/azext_aks_preview/_helpers.py
@@ -16,6 +16,12 @@ from azext_aks_preview._client_factory import (
     get_mc_snapshots_client,
     get_nodepool_snapshots_client,
 )
+
+from azext_aks_preview._consts import (
+    ADDONS,
+    CONST_MONITORING_ADDON_NAME
+)
+
 from azure.cli.command_modules.acs._helpers import map_azure_error_to_cli_error
 from azure.cli.command_modules.acs._validators import extract_comma_separated_string
 from azure.cli.core.azclierror import (
@@ -346,3 +352,18 @@ def check_is_azure_cli_core_editable_installed():
     except Exception as ex:  # pylint: disable=broad-except
         logger.debug("failed to check if azure-cli-core is installed as editable: %s", ex)
     return False
+
+
+def is_monitoring_addon_enabled(addons, instance):
+        is_monitoring_addon = False
+        addon_args = addons.split(',')
+        for addon_arg in addon_args:
+            if addon_arg in ADDONS:
+                addon = ADDONS[addon_arg]
+                if addon == CONST_MONITORING_ADDON_NAME:
+                    is_monitoring_addon = True
+                    break
+
+        addon_profiles = instance.addon_profiles or {}
+        return is_monitoring_addon and CONST_MONITORING_ADDON_NAME in addon_profiles and addon_profiles[
+            CONST_MONITORING_ADDON_NAME].enabled
